the background color of flagged tiles is darker
reason: when delta mode is off, your tile-counting has to include both covered and flagged tiles, so it makes sense to give them the same background color. also, the blue highlight is much harder to see on tiles with light backgrounds, and with delta mode off it's important that you can see it on flags.